% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthomap.R
\name{orthomap}
\alias{orthomap}
\title{Create a World map in orthographic projection}
\usage{
orthomap(query = NULL, centre = c(0, 0), border = NA, fill = "#909090",
  grid = TRUE, grid.color = "#969696", grid.type = 1, grid.size = 0.25,
  ...)
}
\arguments{
\item{query}{The name of a country to center the map on.}

\item{centre}{Latitude and longitude of the map center. Ignored if query is not null.}

\item{border}{Color of polygons border.}

\item{fill}{Color of polygons.}

\item{grid}{If TRUE, grid (i.e. graticules) are added to the map.}

\item{grid.color}{Color of the grid.}

\item{grid.type}{Type of grid (see argument \code{lty}).}

\item{grid.size}{Size of grid (see argument \code{lwd}).}
}
\value{
A \code{SpatialPolygons} of world countries in orthographic projection without aberrations.
}
\description{
This function corrects a bug (some polygons are not correctly projected) in
the function \code{maps:map()} when it is used in orthographic projection. The
function \code{orthomap()} also returns projected polygons in the
\code{sp:SpatialPolygons} format.
}
\examples{
### Map centered on Sao Tome and Principe
world <- orthomap(query = "Sao Tome and Principe")

### Class of the object returned
class(world)

### Map centered on North Pole
world <- orthomap(centre = c(90, 0))

### Map with user-defined graphical parameters
world <- orthomap(centre = c(90, 0), mar = rep(0, 4), bg = "black")
}
\seealso{
\code{\link[maps]{map}}
}
