% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthomap.R
\name{orthomap}
\alias{orthomap}
\title{Create a World map in orthographic projection}
\usage{
orthomap(query = NULL, centre = c(0, 0), border.color = NA, border.type  = 1,
  border.size  = 0.25, fill = "#909090", globe = "transparent", grid = TRUE,
  nx = 10, ny = 10, grid.color = "#969696", grid.type = 1, grid.size = 0.25,
  ...)
}
\arguments{
\item{query}{The name of a country to center the map on.}

\item{centre}{Latitude and longitude of the map center. Ignored if query is not null.}

\item{border.color}{Color of polygons border.}

\item{border.type}{Type of polygons border line (see argument \code{lty}).}

\item{border.size}{Size of polygons border line (see argument \code{lwd}).}

\item{fill}{Color of polygons.}

\item{globe}{Color of globe background.}

\item{grid}{If TRUE, grid (i.e. graticules) are added to the map.}

\item{nx}{Number of longitude lines.}

\item{ny}{Number of latitude lines.}

\item{grid.color}{Color of the grid.}

\item{grid.type}{Type of grid (see argument \code{lty}).}

\item{grid.size}{Size of grid (see argument \code{lwd}).}

\item{...}{Other graphical parameters (see \code{par}).}
}
\value{
A \code{SpatialPolygons} of world countries in orthographic projection without aberrations.
}
\description{
This function corrects a bug (some polygons are not correctly projected) in
the function \code{maps:map()} when it is used in orthographic projection. The
function \code{orthomap()} also returns projected polygons in the
\code{sp:SpatialPolygons} format.
}
\seealso{
\code{\link[maps]{map}}
}
\examples{
# Default settings
world <- orthomap()

class(world)
length(world)
plot(world)

# Center map based on country name
world <- orthomap(query = "France")

# Center map based on coordinates
world <- orthomap(centre = c(45, 125))

# Customize map
world <- orthomap(
  centre       = c(45, 125),
  border.color = "#dddddd",
  border.type  = 1,
  border.size  = 0.25,
  fill         = c("#0868ac", "#016c59", "#b30000"),
  globe        = "#b3d0d1",
  grid         = TRUE,
  nx           = 20,
  ny           = 20,
  grid.color   = "#969696",
  grid.type    = 1,
  grid.size    = 0.25,
  mar          = rep(0, 4),
  bg           = "#dddddd"
)
}
